Mannequin step sounds changed (two different ones now).
I could not get the mannequins to work in the game, so I did not fiddle too much around with it (in case I would f'ck something up). All I did was change the code and sound files from the previous sound implementation for the "Step" sound in the Mannequine script.